+++ libraries/printHelpers/CHANGELOG.md
@@ -6,13 +6,18 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 
+## [0.5.2] - 2026-05-13
+- fix #30, handling infinity (also in printHelpersMT.h)
+- add infinity + nan test in build-CI
+- add print_double_test.ino to find max and min double.
+- minor edits
+
 ## [0.5.1] - 2025-12-29
 - fix #26, add units() e.g. to replace E+xx => mVolt 
 - fix #27, add rightAlign to eng() notation.
 - update examples
 - update GitHub actions
 - minor edits
-
 
 ## [0.5.0] - 2025-06-30
 - fix #24, add printHelpersMT.h,
@@ -123,4 +128,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [0.1.0] -  2018-01-21
 - initial version
 
+----
+
+## [0.0.0] - 2013-05-12
+
+https://forum.arduino.cc/t/proposed-update-for-the-printfloat-code-of-print-cpp/162240